+++ spython/main/base/__init__.py
@@ -30,6 +30,7 @@ from .command import ( generate_bind_list, init_command, run_command )
 from .flags import parse_verbosity
 from .sutils import ( get_uri, load, setenv, get_filename )
 from .logger import ( println,  init_level )
+from .generate import RobotNamer
 
 class Client:
 
@@ -64,12 +65,13 @@ class Client:
             return version
 
 
-    def check_install(self):
+    def _check_install(self):
         '''ensure that singularity is installed, and exit if not.
         '''
         if check_install() is not True:
             bot.error("Cannot find Singularity! Is it installed?")
             sys.exit(1)
+
 
 
 # Image Utils
@@ -87,3 +89,4 @@ Client._run_command = run_command
 Client._parse_verbosity = parse_verbosity
 Client._println = println
 Client._init_level = init_level
+Client.RobotNamer = RobotNamer()