+++ spython/main/help.py
@@ -33,7 +33,8 @@ def help(self, command=None):
         command: the command to get help for, if none, prints general help
 
     '''
-    self.check_install()
+    from spython.utils import check_install
+    check_install()
 
     cmd = ['singularity','--help']
     if command != None: