+++ spython/main/run.py
@@ -49,13 +49,18 @@ def run(self,
         stream: if True, return <generator> for the user to run
 
     '''
+    from spython.utils import check_install
+    check_install()
 
-    self.check_install()
     cmd = self._init_command('run')
 
     # No image provided, default to use the client's loaded image
     if image is None:
         image = self._get_uri()
+
+    # If an instance is provided, grab it's name
+    if isinstance(image, self.instance):
+        image = image.get_uri()
 
     # Does the user want to use bind paths option?
     if bind is not None: