+++ spython/utils/__init__.py
@@ -1,2 +1,16 @@
-from .fileio import *
-from .terminal import *
+from .fileio import ( 
+    mkdir_p, 
+    write_file, 
+    write_json,
+    read_file, 
+    read_json
+)
+
+from .terminal import ( 
+    check_install, 
+    get_installdir,
+    stream_command,
+    run_command,
+    format_container_name,
+    remove_uri
+)