+++ spython/utils/terminal.py
@@ -28,9 +28,9 @@ import subprocess
 import sys
 
 
-##########################################################################
+################################################################################
 # Local commands and requests
-##########################################################################
+################################################################################
 
 
 def check_install(software='singularity', quiet=True):
@@ -95,7 +95,12 @@ def stream_command(cmd, no_newline_regexp="Progess", sudo=False):
         raise subprocess.CalledProcessError(return_code, cmd)
 
 
-def run_command(cmd, sudo=False, capture=True, no_newline_regexp="Progess", quiet=False):
+def run_command(cmd, 
+                sudo=False,
+                capture=True,
+                no_newline_regexp="Progess",
+                quiet=False):
+
     '''run_command uses subprocess to send a command to the terminal. If
        capture is True, we use the parent stdout, so the progress bar (and
        other commands of interest) are piped to the user. This means we 
@@ -147,7 +152,12 @@ def run_command(cmd, sudo=False, capture=True, no_newline_regexp="Progess", quie
     return output
 
 
+################################################################################
+# Parsing and Formatting
+################################################################################
 
+
+         
 
 def format_container_name(name, special_characters=None):
     '''format_container_name will take a name supplied by the user,