+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The client here will eventually be released as "spython" (and eventually to
 singularity on pypi), and the versions here will coincide with these releases.
 
 ## [master](https://github.com/singularityhub/singularity-cli/tree/master)
+ - adding support and documentation for container instances (0.0.38)
  - fixing bug with recipe Dockerfile conversion (0.0.37)
  - typo in pypi PACKAGE_URL (0.0.36)
  - respecting Client "quiet" attribute in run_command  (0.0.34/0.0.35)